+++ src/Stormancer.Plugins/Party/Stormancer.Server.Plugins.Party/Changelog.rst
@@ -16,6 +16,7 @@ Changed
 *******
 - Updated dependency to Users to support new abstraction lib.
 - Remove current gamesession in party when the gamesession scene shuts down.
+- If party does not exist, CreateConnectionToken returns 'notFound'
 
 Fixed
 *****